lib-kmm-foundation/src/androidMain/kotlin/com/hadisatrio/libs/android/foundation/event/ExecutorDispatchingEventSink.kt
/*
 * Copyright (C) 2022 Hadi Satrio
 *
 * This program is free software: you can redistribute it and/or modify
 * it under the terms of the GNU General Public License as published by
 * the Free Software Foundation, either version 3 of the License, or
 * (at your option) any later version.
 *
 * This program is distributed in the hope that it will be useful,
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 * GNU General Public License for more details.
 *
 * You should have received a copy of the GNU General Public License
 * along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */

package com.hadisatrio.libs.android.foundation.event

import com.hadisatrio.libs.kotlin.foundation.event.Event
import com.hadisatrio.libs.kotlin.foundation.event.EventSink
import java.util.concurrent.Executor

class ExecutorDispatchingEventSink(
    private val executor: Executor,
    private val origin: EventSink
) : EventSink {

    override fun sink(event: Event) {
        executor.execute { origin.sink(event) }
    }
}
